// iglesia-bautismo-frontend/src/types/Certificado.ts
export interface Certificado {
  id: number;
  nombreSuscribe: string;
  libroBautizo: string;
  folioBautizo: string;
  numeroArchivo: string;
  nombreBautizado: string;
  diaNacimiento: number | null;
  mesNacimiento: string;
  anoNacimiento: number | null;
  lugarNacimiento: string;
  nombrePadre: string;
  nombreMadre: string;
  libroNacimiento: string;
  folioNacimiento: string;
  archivoNacimiento: string;
  anoArchivo: string;
  oficialia: string;
  diaBautismo: number | null;
  mesBautismo: string;
  anoBautismo: number | null;
  padrino: string;
  madrina: string;
  ministroSacramento: string;
  notasMarginales: string;
  diaEmision: number | null;
  mesEmision: string;
  anoEmision: number | null;
  genero: string;
}